Hashing implementation (#73)
