+++ src/discobase/cogs/utility.py
@@ -11,7 +11,6 @@ class Utility(commands.Cog):
     def __init__(self, bot: commands.Bot) -> None:
         self.bot = bot
 
-
     @app_commands.command(description="Insert new data into a table.")
     @app_commands.describe(
         table="Choose the table you want to insert the data into.",
