+++ src/discobase/table.py
@@ -1,14 +1,17 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, ClassVar, Optional, Set
+from typing import (TYPE_CHECKING, Any, ClassVar, Literal, Optional, Set,
+                    overload)
 
-from pydantic import BaseModel
-from typing_extensions import Self
+from pydantic import BaseModel, ConfigDict
+from typing_extensions import Self, Unpack
 
 if TYPE_CHECKING:
     from .database import Database
 
-from .exceptions import DatabaseTableError, NotConnectedError
+from ._cursor import TableCursor
+from .exceptions import (DatabaseLookupError, DatabaseStorageError,
+                         DatabaseTableError, NotConnectedError)
 
 __all__ = ("Table",)
 
@@ -16,14 +19,27 @@ __all__ = ("Table",)
 # Note that we can't use 3.10+ type[] syntax
 # here, since Pydantic can't handle it
 class Table(BaseModel):
-    __disco_database__: ClassVar[Optional[Database]] = None
+    __disco_database__: ClassVar[Optional[Database]]
     """Attached `Database` object. Set by the `table()` decorator."""
-    __disco_keys__: ClassVar[Set[str]] = set()
+    __disco_cursor__: ClassVar[Optional[TableCursor]]
+    """Internal table cursor, set at initialization time."""
+    __disco_keys__: ClassVar[Set[str]]
     """All keys of the table, this may not change once set by `table()`."""
-    __disco_name__: ClassVar[str] = "_notset"
+    __disco_name__: ClassVar[str]
     """Internal name of the table. Set by the `table()` decorator."""
-    __disco_ready__: ClassVar[bool] = False
-    """Whether the `Table` object has it's database channels set up."""
+    __disco_id__: int = -1
+    """Message ID of the record. This is only present if it was saved."""
+
+    def __init__(self, /, **data: Any) -> None:
+        super().__init__(**data)
+        self.__disco_id__ = -1
+
+    def __init_subclass__(cls, **kwargs: Unpack[ConfigDict]) -> None:
+        super().__init_subclass__(**kwargs)
+        cls.__disco_database__ = None
+        cls.__disco_cursor__ = None
+        cls.__disco_keys__ = set()
+        cls.__disco_name__ = "_notset"
 
     @classmethod
     def _ensure_db(cls) -> None:
@@ -37,14 +53,14 @@ class Table(BaseModel):
                 "database is not connected! did you forget to open it?"
             )
 
-        if not cls.__disco_ready__:
+        if not cls.__disco_cursor__:
             raise DatabaseTableError(
                 f"{cls.__name__} is not ready, you might want to add a call to build_tables()",  # noqa
             )
 
     async def save(self) -> None:
         """
-        Commit the current object to the database.
+        Save the entry to the database as a new record.
 
         Example:
             ```py
@@ -62,15 +78,55 @@ class Table(BaseModel):
             ```
         """
         self._ensure_db()
-        assert self.__disco_database__
-        await self.__disco_database__._add_record(self)
+        assert self.__disco_cursor__
+
+        if self.__disco_id__ != -1:
+            raise DatabaseStorageError(
+                "this entry has already been written, did you mean to call update()?",  # noqa
+            )
+        msg = await self.__disco_cursor__.add_record(self)
+        self.__disco_id__ = msg.id
+
+    async def update(self) -> None:
+        """
+        Update the entry in-place.
+
+        Example:
+            ```py
+            import discobase
+
+            db = discobase.Database("My database")
+
+            @db.table
+            class User(discobase.Table):
+                name: str
+                password: str
+
+            # Using top-level await for this example
+            user = await User.find_unique(name="Peter", password="foobar")
+            user.password = str(hash(password))
+            await user.update()
+            ```
+        """
+
+        self._ensure_db()
+        assert self.__disco_cursor__
+        if self.__disco_id__ == -1:
+            raise DatabaseStorageError(
+                "this entry has not been written, did you mean to call save()?",  # noqa
+            )
+        await self.__disco_cursor__.update_record(self)
+
+    async def commit(self) -> None: ...
 
     @classmethod
     async def find(cls, **kwargs: Any) -> list[Self]:
         """
-        Find a list of instances of the schema type.
+        Find a list of entries in the database.
+
         Args:
             **kwargs: Values to search for. These should be keys in the schema.
+
         Example:
             ```py
             import discobase
@@ -84,8 +140,60 @@ class Table(BaseModel):
             ```
         """
         cls._ensure_db()
-        assert cls.__disco_database__
-        return await cls.__disco_database__._find_records(
-            cls.__disco_name__,
+        assert cls.__disco_cursor__
+        return await cls.__disco_cursor__.find_records(
+            cls,
             kwargs,
         )
+
+    @classmethod
+    @overload
+    async def find_unique(
+        cls,
+        *,
+        strict: Literal[True] = True,
+        **kwargs: Any,
+    ) -> Self: ...
+
+    @classmethod
+    @overload
+    async def find_unique(
+        cls,
+        *,
+        strict: Literal[False] = False,
+        **kwargs: Any,
+    ) -> Self | None: ...
+
+    @classmethod
+    async def find_unique(
+        cls,
+        *,
+        strict: bool = True,
+        **kwargs: Any,
+    ) -> Self | None:
+        """
+        Find a unique entry in the database.
+
+        Args:
+            **kwargs: Values to search for. These should be keys in the schema.
+        """
+
+        if not kwargs:
+            raise ValueError("a query must be passed to find_unique")
+
+        values: list[Self] = await cls.find(**kwargs)
+
+        if not len(values):
+            if strict:
+                raise DatabaseLookupError(
+                    f"no entry found with query {kwargs}",
+                )
+
+            return None
+
+        if strict and (1 < len(values)):
+            raise DatabaseLookupError(
+                "more than one entry was found with find_unique"
+            )
+
+        return values[0]
